backend/requirements.txt
# Backend dependencies for Colombia News
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Scraping
playwright==1.41.0
newspaper3k==0.2.8
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
httpx==0.25.2

# AI & Embeddings
openai==1.10.0
anthropic==0.18.1
langchain==0.1.5
langchain-openai==0.0.5
langchain-anthropic==0.1.1

# Vector & Database
supabase==2.3.4
pgvector==0.2.4
psycopg2-binary==2.9.9

# Clustering & ML
scikit-learn==1.4.0
numpy==1.26.3

# Task Queue
celery==5.3.6
redis==5.0.1

# Utils
python-dateutil==2.8.2
pytz==2024.1
